It resolves
